Alright sit down and buckle up cuz this is gonna be a wild ride kiddos. This push will hopefully fufill kevins request. Thank. You. Kevin. this has been sooo much fun.
Alright, essentialy this code cuts out the middle man of reseting the 
robot position after it knocks the mineral. So it runs over the mineral 
and keeps going for an unknown distance (distanceForward), once it is 
far enough away from the other minerals it does an unknown spot turn so 
that it is perpendicular with the wall and the dropper (which should be 
on the left of the chassis) is facing the depot. The robot then drives 
forward so that it is nearly kissing the wall (distanceForward2) then it 
shimmys so that the dropper is at the edge of the depot 
(distanceShimmy2) then it drops the marker. Then it shimmys along the 
wall to the crater. ez peasy lemon squeezy.

All of these values are untested and therefore unknown. Your mission 
should you choose to accept it (theres no choice its an order) is to 
test values until you get it right. You can look at previous methods to 
get a starting idea of what values for what distance, but it will take 
time to refine them. 

currently this new code is commented out, when testing uncomment it and 
then comment out the code below it (put the closing one below 
"penetrates the crater") 

have fun mindlessly testing values and happy trails friend